requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
langchain_openai
langchain-community
faiss-cpu